+++ src/routes/posts/+page.js
@@ -1,0 +1,9 @@
+// @ts-ignore
+export const load = async ({ fetch }) => {
+  const posts = await fetch('/api/posts.json')
+  const allPosts = await posts.json()
+
+  return {
+  posts: allPosts
+}
+}
