+++ src/routes/posts/[page=int]/+page.svelte
@@ -1,18 +1,6 @@
-<script context="module">
-    // @ts-ignore
-    export const load = async ({ fetch }) => {
-      const posts = await fetch('/api/posts.json')
-      const allPosts = await posts.json()
-    
-      return {
-        props: {
-          posts: allPosts
-        }
-      }
-    }
-</script>
-
 <script lang="ts">
+    throw new Error("@migration task: Add data prop (https://github.com/sveltejs/kit/discussions/5774#discussioncomment-3292707)");
+
     import { page } from '$app/stores';
     import ShortPost from '$lib/components/ShortPost.svelte';
     import Pagination from '$lib/components/Pagination.svelte';
