+++ src/routes/posts/[post]/+page.svelte
@@ -1,22 +1,6 @@
-<script context="module">
-  // @ts-ignore
-  export const load = async ({ params, fetch }) => {
-    const currentPath = params.post
-    const response = await fetch('/api/posts.json')
-    const posts = await response.json()
-
-    const matchingPosts = posts.filter(post => post.meta.path ? post.meta.path == currentPath : false)
-
-    return {
-      props: {
-        posts: matchingPosts,
-        path: currentPath
-      }
-    }
-  }
-</script>
-
 <script lang="ts">
+  throw new Error("@migration task: Add data prop (https://github.com/sveltejs/kit/discussions/5774#discussioncomment-3292707)");
+
   import Tag from '$lib/components/Tag.svelte';
 
   export let posts;
